Set fixed size on scancode->ascii array
Thus every element without a specific value will be initialized to 0.
If we don't specify a size the compiler will make the smallest fit,
and thus reads with an offset higher than the highest specified index
will go out of bound and read invalid memory! We want it to read 0!